Col spreads props into first child div
To toggle on and off different sections of a page, Optimizely needs to
be able to find certain items using Jquery. In Gatsby, we added a `data`
prop to certain form elements. This prop was getting swallowed by `Col`.
This PR stops `data-*` (and other props, including, but not limited to,
ID) from disappearing into the abyss.