Add background block and gray box components
Since the gray box is a central part of the new branding, we made it
reusable here. A few caveats: the `GrayBox` is an extension of `Layout`,
and the `verticalPadding` and `leftOffset` variants use `StyledWrapper`
which can behave strangely with advanced usage. These features work best
when the props for offset/ padding are not going to change at runtime.