Add GrayBox to Layout
The current GrayBox component works great for it's intended purpose:
being a gray box that holds some content. However, sometimes we need a
gray box that is merely a background and doesn't care if there is
content inside. LayoutGrayBox is that. When placed inside a layout it
serves as a background to that layout and doesn't affect the other
columns within that layout