Update GrayBox naming
Renamed GreyBox -> GrayBox
Moved atoms/GrayBox -> BlockHeader/GrayBox

The GrayBox that is a container being used by BlockHeader is now moved
into the BlockHeader folder and will be used as a private component. We
shouldn't need to use this in the RCL moving forward.
GrayBox, a Column that serves as a background to the Layout Component,
is now the go to GrayBox and is exposed in the RCL