Bug - Wrong Date on FullDateField when 1st of Month
The `FullDateField` was displaying the previous day when passed a date
that was the first of the month. This has to do with how `new Date()`
works and does some auto-adjusting behind the scenes. It is safer to
pass the date values directly into the constructor i.e `new Date(2018,
10, 29)`