MOR-9: Typography adjustments
- Update the `A` and `B` fonts in our typography. Some
of the styles are consolidated into one (eg. A6 and A7). In addition to
this consolidation, the `B` font will now be the serif font `Lora`.
- In addition to the style updates, I also updated the styleguidist
documentation a bit for the Branding section; now, the Typography section
just points right to the documentation for the Text component, and the
Spacer section points to the Spacer component.